refactor: isolate spill behind sidecar plugin
