+++ website/src/content/docs/docs/concepts/tape-and-context.mdx
@@ -50,7 +50,7 @@ The default `provide_tape_store` returns a `FileTapeStore` rooted at `~/.bub/tap
 
 ### Spill sidecar
 
-The builtin spill plugin mounts a `SpillStore` through `provide_tape_sidecars`. Large string tool results are stored in a sibling tape named `<session-tape>__spill`. The sidecar uses the same `TapeStore` as the session tape, so existing storage plugins do not need a spill-specific interface. Bub writes UTF-8-safe chunks followed by a manifest; the manifest is the completion marker for the stored result.
+The builtin spill plugin mounts a `SpillStore` through `provide_tape_sidecars` and registers `spill.read`. It owns tool-result handling without exposing spill-specific types to the core. Large results are stored in a sibling tape named `<session-tape>__spill`. The sidecar uses the same `TapeStore` as the session tape, so existing storage plugins do not need a spill-specific interface. Bub writes UTF-8-safe chunks followed by a manifest; the manifest is the completion marker for the stored result.
 
 The main tape keeps a bounded preview and an opaque handle instead of the complete result. The `spill.read` tool reads a bounded page by handle and cursor, including pages counted from the end. Content explicitly returned by `spill.read` is recorded as a normal bounded tool result.
